refactor(bot): move voice-channel-status service into handlers/player
VoiceChannelStatusService sat bare in top-level services/ alongside
genuinely cross-cutting services (HeartbeatService, WeeklyDigestService,
guildMembershipService), but all 3 of its consumers are in
handlers/player/ - the discord-player playback engine, not any
cross-cutting concern (#1974). Moved it there as voiceChannelStatusService.ts
(matching that directory's camelCase file convention) and updated the
3 import sites plus the two spec files' mock/import paths.

Left services/musicRecommendation/feedbackService.ts where it is: it's
already namespaced under its own musicRecommendation/ subdirectory
(not bare top-level services/ pollution), which is the alternative
home #1974 itself names as acceptable pending #1968's broader
utils/music/ reorg - moving it now risks double-churn.
