+++ packages/bot/src/handlers/player/trackHandlers.ts
@@ -14,7 +14,7 @@ import {
 } from './trackNowPlaying'
 import { musicWatchdogService } from '../../services/musicManagement/watchdog'
 import { musicSessionSnapshotService } from '../../services/musicRecommendation/sessionSnapshots'
-import * as voiceStatus from '../../services/VoiceChannelStatusService'
+import * as voiceStatus from './voiceChannelStatusService'
 import {
     scheduleIdleDisconnect,
     clearIdleTimer,
